finish analyzer cleanup follow-up
