+++ scripts/provider_ui/provider_admin_bootstrap.mjs
@@ -352,9 +352,5 @@ function reportCliError(error) {
 }
 
 if (isCliEntrypoint(import.meta.url)) {
-  try {
-    await main();
-  } catch (error) {
-    reportCliError(error);
-  }
+  main().catch(reportCliError);
 }
