+++ scripts/quality/render_codex_prompt.py
@@ -55,15 +55,9 @@ def _repo_contract_lines(profile: dict) -> List[str]:
         f"- Codex auth lane: `{profile.get('codex_auth_lane', 'chatgpt-account')}`",
         f"- Provider UI mode: `{provider_ui_mode}`",
         f"- Codex environment mode: `{codex_environment.get('mode', 'automatic')}`",
-        (
-            "- Codex environment verify command: "
-            f"`{codex_environment.get('verify_command', profile['verify_command'])}`"
-        ),
+        f"- Codex environment verify command: `{codex_environment.get('verify_command', profile['verify_command'])}`",
         f"- Codex auth file: `{codex_auth_file}`",
-        (
-            "- Codex environment network profile: "
-            f"`{codex_environment.get('network_profile', 'unrestricted')}`"
-        ),
+        f"- Codex environment network profile: `{codex_environment.get('network_profile', 'unrestricted')}`",
         f"- Codex environment methods: `{codex_environment.get('methods', 'all')}`",
         f"- Codex runner labels: `{runner_labels}`",
         f"- Default branch: `{profile['default_branch']}`",
