+++ tests/test_check_sonar_zero.py
@@ -124,128 +124,6 @@ class SonarZeroTests(unittest.TestCase):
                 {"paging": {"total": 0}},
             )
 
-    def test_load_helpers_collect_open_issues_quality_gate_and_findings(self) -> None:
-        """Cover load helpers collect open issues quality gate and findings."""
-        args = Namespace(
-            project_key="project", branch="", pull_request="5", policy_mode="zero"
-        )
-        branch_args = Namespace(
-            project_key="project", branch="main", pull_request="", policy_mode="zero"
-        )
-        self._assert_sonar_request_round_trip(
-            {
-                "args": args,
-                "responses": [
-                    {"paging": {"total": 1}},
-                    {"projectStatus": {"status": "ERROR"}},
-                ],
-                "expected_open_issues": 1,
-                "expected_quality_gate": "ERROR",
-                "expected_query": "pullRequest=5",
-            }
-        )
-        self._assert_sonar_request_round_trip(
-            {
-                "args": branch_args,
-                "responses": [
-                    {"paging": {"total": 0}},
-                    {"projectStatus": {"status": "OK"}},
-                ],
-                "expected_open_issues": 0,
-                "expected_quality_gate": "OK",
-                "expected_query": "branch=main",
-            }
-        )
-
-        with patch.object(
-            check_sonar_zero, "_load_open_issues", return_value=2
-        ), patch.object(check_sonar_zero, "_load_quality_gate", return_value="ERROR"):
-            open_issues, quality_gate, findings = check_sonar_zero._load_sonar_findings(
-                args, "auth"
-            )
-        self.assertEqual(open_issues, 2)
-        self.assertEqual(quality_gate, "ERROR")
-        self.assertIn("Sonar reports 2 open issues", findings[0])
-        self.assertIn("Sonar quality gate status is ERROR", findings[1])
-
-        with patch.object(
-            check_sonar_zero, "_load_open_issues", return_value=0
-        ), patch.object(check_sonar_zero, "_load_quality_gate", return_value="OK"):
-            self.assertEqual(
-                check_sonar_zero._load_sonar_findings(args, "auth"), (0, "OK", [])
-            )
-
-        ratchet_args = Namespace(
-            project_key="project", branch="", pull_request="5", policy_mode="ratchet"
-        )
-        with patch.object(
-            check_sonar_zero, "_load_open_issues", return_value=6
-        ), patch.object(check_sonar_zero, "_load_quality_gate", return_value="OK"):
-            self.assertEqual(
-                check_sonar_zero._load_sonar_findings(ratchet_args, "auth"),
-                (6, "OK", []),
-            )
-
-    def test_revision_helpers_and_pending_message_cover_scoped_paths(self) -> None:
-        """Cover revision helpers and pending message cover scoped paths."""
-        args = Namespace(
-            project_key="project", branch="main", pull_request="", sha="targetsha"
-        )
-        self._assert_revision_lookup(
-            {
-                "args": args,
-                "payload": {"branches": [{"name": "main", "commit": {"sha": "oldsha"}}]},
-                "expected_revision": "oldsha",
-                "expected_pending_message": (
-                    "Sonar analysis for branch main is still on oldsha "
-                    "(waiting for targetsha)."
-                ),
-                "revision_loader": check_sonar_zero._load_branch_analysis_revision,
-            }
-        )
-        self._assert_revision_lookup(
-            {
-                "args": args,
-                "payload": {"branches": [{"name": "other", "commit": {"sha": "oldsha"}}]},
-                "expected_revision": "",
-                "expected_pending_message": "Sonar analysis for branch main is not available yet.",
-                "revision_loader": check_sonar_zero._load_branch_analysis_revision,
-            }
-        )
-
-        pr_args = Namespace(
-            project_key="project", branch="", pull_request="5", sha="targetsha"
-        )
-        self._assert_revision_lookup(
-            {
-                "args": pr_args,
-                "payload": {
-                    "pullRequests": [{"key": "5", "commit": {"sha": "targetsha"}}]
-                },
-                "expected_revision": "targetsha",
-                "expected_pending_message": None,
-                "revision_loader": check_sonar_zero._load_pull_request_analysis_revision,
-            }
-        )
-        self._assert_revision_lookup(
-            {
-                "args": pr_args,
-                "payload": {
-                    "pullRequests": [{"key": "other", "commit": {"sha": "oldsha"}}]
-                },
-                "expected_revision": "",
-                "expected_pending_message": "Sonar analysis for pull request 5 is not available yet.",
-                "revision_loader": check_sonar_zero._load_pull_request_analysis_revision,
-            }
-        )
-
-        self.assertIsNone(
-            check_sonar_zero._scoped_analysis_pending_message(
-                Namespace(project_key="project", branch="", pull_request="", sha=""),
-                "auth",
-            )
-        )
-
     def test_retry_waits_for_pr_scoped_findings_to_settle(self) -> None:
         """Cover retry waits for pr scoped findings to settle."""
         args = argparse.Namespace(branch="", pull_request="5")
@@ -510,105 +388,3 @@ class SonarZeroTests(unittest.TestCase):
 
         self.assertEqual((open_issues, quality_gate, findings), (0, "OK", []))
         self.assertEqual(attempts, [1, 2, 3, 4, 5, 6, 7, 8])
-
-    def test_main_handles_missing_token_success_and_report_failures(self) -> None:
-        """Cover main handles missing token success and report failures."""
-        args = Namespace(
-            project_key="Prekzursil_quality-zero-platform",
-            token=str(),
-            branch="",
-            pull_request="5",
-            out_json="sonar-zero/sonar.json",
-            out_md="sonar-zero/sonar.md",
-        )
-
-        with patch.dict("os.environ", {}, clear=True):
-            self._assert_main_result(
-                {
-                    "args": args,
-                    "expected_code": 1,
-                    "expected_findings": ["SONAR_TOKEN is missing."],
-                }
-            )
-
-        success_args = Namespace(**{**args.__dict__, "token": "provided-token"})
-        self._assert_main_result(
-            {
-                "args": success_args,
-                "load_result": (0, "OK", []),
-                "expected_code": 0,
-                "expected_status": "pass",
-            }
-        )
-        self._assert_main_result(
-            {
-                "args": success_args,
-                "load_side_effect": RuntimeError("provider timeout"),
-                "expected_code": 1,
-                "expected_findings": ["Sonar API request failed: provider timeout"],
-            }
-        )
-        self._assert_main_result(
-            {
-                "args": success_args,
-                "load_result": (0, "OK", []),
-                "write_report_result": 4,
-                "expected_code": 4,
-                "expected_status": "pass",
-            }
-        )
-
-        audit_args = Namespace(**{**success_args.__dict__, "policy_mode": "audit"})
-        self._assert_main_result(
-            {
-                "args": audit_args,
-                "load_result": (
-                    3,
-                    "ERROR",
-                    ["Sonar reports 3 open issues (expected 0)."],
-                ),
-                "expected_code": 0,
-                "expected_status": "pass",
-            }
-        )
-
-    def test_parse_args_render_markdown_and_script_entrypoint(self) -> None:
-        """Cover parse args render markdown and script entrypoint."""
-        with patch.object(
-            sys, "argv", ["check_sonar_zero.py", "--project-key", "project"]
-        ):
-            args = check_sonar_zero._parse_args()
-        self.assertEqual(args.project_key, "project")
-        markdown = check_sonar_zero._render_md(
-            {
-                "status": "pass",
-                "project_key": "project",
-                "open_issues": 0,
-                "quality_gate": "OK",
-                "timestamp_utc": "2026-03-15T00:00:00+00:00",
-                "findings": [],
-            }
-        )
-        self.assertIn("- None", markdown)
-
-        script_path = Path("scripts/quality/check_sonar_zero.py").resolve()
-        root_text = str(Path.cwd().resolve())
-        trimmed_sys_path = [item for item in sys.path if item != root_text]
-        with tempfile.TemporaryDirectory() as tmp, patch.dict(
-            "os.environ", {}, clear=True
-        ), patch.object(
-            sys,
-            "argv",
-            [str(script_path), "--project-key", "Prekzursil_quality-zero-platform"],
-        ), patch.object(
-            sys, "path", trimmed_sys_path[:]
-        ):
-            cwd = Path(tmp)
-            previous = Path.cwd()
-            os.chdir(cwd)
-            try:
-                with self.assertRaises(SystemExit) as result:
-                    runpy.run_path(str(script_path), run_name="__main__")
-            finally:
-                os.chdir(previous)
-        self.assertEqual(result.exception.code, 1)
