+++ tests/test_coverage_assert.py
@@ -170,147 +170,6 @@ class CoverageAssertTests(unittest.TestCase):
             "build/_deps/googletest-src/googletest/src/gtest.cc", lcov_sources
         )
 
-    def test_evaluate_flags_missing_required_sources_and_honors_threshold(self) -> None:
-        """Cover evaluate flags missing required sources and honors threshold."""
-        stats = [CoverageStats(name="python", path="coverage.xml", covered=9, total=10)]
-
-        status, findings = evaluate(
-            stats,
-            CoverageEvaluationRequest(
-                min_percent=100.0,
-                required_sources=["app/main.py"],
-                reported_sources={"tests/test_main.py"},
-            ),
-        )
-
-        self.assertEqual(status, "fail")
-        self.assertTrue(any("coverage below 100.00%" in item for item in findings))
-        self.assertTrue(
-            any(
-                "missing required source path: app/main.py" in item for item in findings
-            )
-        )
-        self.assertTrue(any("tests/ paths" in item for item in findings))
-
-        ok_status, ok_findings = evaluate(
-            stats,
-            CoverageEvaluationRequest(
-                min_percent=90.0,
-                required_sources=["app/main.py"],
-                reported_sources={"app/main.py"},
-            ),
-        )
-
-        self.assertEqual(ok_status, "pass")
-        self.assertEqual(ok_findings, [])
-
-        branch_stats = [
-            CoverageStats(
-                name="python",
-                path="coverage.xml",
-                covered=10,
-                total=10,
-                branch_covered=5,
-                branch_total=10,
-            )
-        ]
-        branch_status, branch_findings = evaluate(
-            branch_stats,
-            CoverageEvaluationRequest(
-                min_percent=100.0,
-                branch_min_percent=80.0,
-                required_sources=["app/main.py"],
-                reported_sources={"app/main.py"},
-            ),
-        )
-        self.assertEqual(branch_status, "fail")
-        self.assertTrue(
-            any("branch coverage below 80.00%" in item for item in branch_findings)
-        )
-
-        missing_branch_status, missing_branch_findings = evaluate(
-            [
-                CoverageStats(
-                    name="no-branch",
-                    path="coverage.xml",
-                    covered=10,
-                    total=10,
-                    branch_covered=0,
-                    branch_total=0,
-                )
-            ],
-            CoverageEvaluationRequest(
-                min_percent=100.0,
-                branch_min_percent=100.0,
-                required_sources=["app/main.py"],
-                reported_sources={"app/main.py"},
-            ),
-        )
-        self.assertEqual(missing_branch_status, "fail")
-        self.assertTrue(
-            any(
-                "branch coverage data missing" in item
-                for item in missing_branch_findings
-            )
-        )
-
-    def test_source_normalization_and_required_source_helpers_cover_edge_cases(
-        self,
-    ) -> None:
-        """Cover source normalization and required source helpers cover edge cases."""
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            with tempfile.TemporaryDirectory() as external_tmp:
-                external_root = Path(external_tmp)
-                external_file = external_root / "external.py"
-                external_file.write_text("print('external')\n", encoding="utf-8")
-                (root / "src").mkdir(parents=True)
-                (root / "src" / "main.py").write_text(
-                    "print('main')\n", encoding="utf-8"
-                )
-                (root / "node_modules").mkdir()
-                (root / "node_modules" / "shim.js").write_text(
-                    "export default true;\n", encoding="utf-8"
-                )
-            with _temporary_cwd(root):
-                workspace_root = Path.cwd().resolve(strict=False).as_posix().rstrip("/")
-                self.assertEqual(
-                    _normalize_source_path(f"{workspace_root}/src/main.py"),
-                    "src/main.py",
-                )
-                self.assertEqual(_normalize_source_path("./"), "")
-                self.assertEqual(
-                    _normalize_source_path("./src//main.py"), "src/main.py"
-                )
-                self.assertEqual(_normalize_source_path("."), "")
-                self.assertEqual(_normalize_source_path(workspace_root), "")
-                self.assertEqual(
-                    _normalize_source_path(external_file.as_posix()),
-                    external_file.resolve(strict=False).as_posix(),
-                )
-                self.assertEqual(_existing_repo_file_candidate(""), "")
-                self.assertEqual(
-                    _existing_repo_file_candidate("repo/src/main.py"), "src/main.py"
-                )
-                self.assertFalse(_should_track_coverage_source(""))
-                self.assertFalse(_should_track_coverage_source("node_modules/shim.js"))
-                self.assertFalse(_matches_required_source("src/main.py", ""))
-                self.assertTrue(_matches_required_source("src/main.py", "src"))
-                self.assertEqual(
-                    _find_missing_required_sources(
-                        {"src/main.py"}, ["", "src", "frontend/app.ts"]
-                    ),
-                    ["frontend/app.ts"],
-                )
-                self.assertTrue(_is_tests_only_report({"tests/test_main.py"}))
-                self.assertEqual(
-                    _required_source_findings({"tests/test_main.py"}, ["src/main.py"]),
-                    [
-                        "coverage inputs only reference tests/ paths; first-party sources are missing.",
-                        "missing required source path: src/main.py",
-                    ],
-                )
-
     def test_collect_inputs_build_payload_and_render_markdown(self) -> None:
         """Cover collect inputs build payload and render markdown."""
         with tempfile.TemporaryDirectory() as tmp:
@@ -506,62 +365,3 @@ class CoverageAssertTests(unittest.TestCase):
         )
 
         self.assertEqual(payload["components"][0]["branch_percent"], 90.0)
-
-    def test_script_entrypoint_handles_import_guard_and_report_failures(self) -> None:
-        """Cover script entrypoint handles import guard and report failures."""
-        with tempfile.TemporaryDirectory() as tmp:
-            root = Path(tmp)
-            xml_path = root / "coverage.xml"
-            xml_path.write_text(
-                '<coverage lines-valid="1" lines-covered="1"><class filename="app/main.py" /></coverage>',
-                encoding="utf-8",
-            )
-            script_path = ROOT / "scripts" / "quality" / "assert_coverage_100.py"
-            root_text = str(ROOT)
-            trimmed_sys_path = [item for item in sys.path if item != root_text]
-
-            with _temporary_cwd(root), patch.object(
-                sys,
-                "argv",
-                [
-                    str(script_path),
-                    "--xml",
-                    f"python={xml_path}",
-                    "--out-json",
-                    "coverage-100/out.json",
-                    "--out-md",
-                    "coverage-100/out.md",
-                ],
-            ), patch.object(sys, "path", trimmed_sys_path[:]), self.assertRaises(
-                SystemExit
-            ) as result:
-                runpy.run_path(str(script_path), run_name="__main__")
-
-            self.assertEqual(result.exception.code, 0)
-            self.assertTrue(
-                any(path.name == "out.json" for path in root.rglob("out.json"))
-            )
-
-        parsed_args = Namespace(
-            xml=[],
-            lcov=[],
-            require_source=[],
-            min_percent=100.0,
-            branch_min_percent="",
-            out_json="coverage-100/coverage.json",
-            out_md="coverage-100/coverage.md",
-        )
-        collected_inputs = (
-            [CoverageStats("python", "coverage.xml", 1, 1)],
-            {"pkg/main.py"},
-        )
-        with (
-            patch.object(assert_coverage_100, "_parse_args", return_value=parsed_args),
-            patch.object(
-                assert_coverage_100,
-                "_collect_coverage_inputs",
-                return_value=collected_inputs,
-            ),
-            patch.object(assert_coverage_100, "write_report", return_value=5),
-        ):
-            self.assertEqual(assert_coverage_100.main(), 5)
