+++ tests/test_quality_rollup_extra.py
@@ -23,27 +23,10 @@ FAKE_GITHUB_CREDENTIAL = "gh-auth-placeholder"
 class QualityRollupExtraTests(unittest.TestCase):
     """Quality Rollup Extra Tests."""
 
-    def test_parse_args_and_context_helpers_cover_default_paths(self) -> None:
-        """Cover parse args and context helpers cover default paths."""
-        with patch(
-            "sys.argv",
-            [
-                "build_quality_rollup.py",
-                "--profile-json",
-                "profile.json",
-                "--repo",
-                "owner/repo",
-                "--sha",
-                "abc",
-                "--artifacts-dir",
-                "artifacts",
-            ],
-        ):
-            args = build_quality_rollup.parse_args()
-        self.assertEqual(args.out_json, build_quality_rollup.DEFAULT_ROLLUP_JSON)
-        self.assertEqual(args.out_md, build_quality_rollup.DEFAULT_ROLLUP_MD)
-
-        contexts = {
+    @staticmethod
+    def _context_fixture():
+        """Return one mixed set of check-run and status contexts."""
+        return {
             "Coverage 100 Gate": {
                 "state": "completed",
                 "conclusion": "success",
@@ -70,6 +53,42 @@ class QualityRollupExtraTests(unittest.TestCase):
                 "source": "status",
             },
         }
+
+    @staticmethod
+    def _rollup_main_args(root: Path, profile_path: Path) -> Namespace:
+        """Build the common CLI namespace for rollup-main tests."""
+        return Namespace(
+            profile_json=str(profile_path),
+            repo="owner/repo",
+            sha="abc123",
+            artifacts_dir=str(root),
+            out_json="quality-rollup/summary.json",
+            out_md="quality-rollup/summary.md",
+        )
+
+    def test_parse_args_uses_default_output_paths(self) -> None:
+        """Cover rollup parse-args defaults."""
+        with patch(
+            "sys.argv",
+            [
+                "build_quality_rollup.py",
+                "--profile-json",
+                "profile.json",
+                "--repo",
+                "owner/repo",
+                "--sha",
+                "abc",
+                "--artifacts-dir",
+                "artifacts",
+            ],
+        ):
+            args = build_quality_rollup.parse_args()
+        self.assertEqual(args.out_json, build_quality_rollup.DEFAULT_ROLLUP_JSON)
+        self.assertEqual(args.out_md, build_quality_rollup.DEFAULT_ROLLUP_MD)
+
+    def test_status_and_lane_detail_helpers_cover_default_paths(self) -> None:
+        """Cover status resolution and lane-detail helpers."""
+        contexts = self._context_fixture()
         self.assertEqual(
             build_quality_rollup._status_from_context("Missing", contexts), "missing"
         )
@@ -104,8 +123,10 @@ class QualityRollupExtraTests(unittest.TestCase):
             build_quality_rollup._lane_detail({"mode": "audit"}), "Mode: audit"
         )
 
-    def test_github_payload_and_load_contexts_cover_non_default_paths(self) -> None:
-        """Cover github payload and load contexts cover non default paths."""
+    def test_github_payload_accepts_check_runs_and_rejects_invalid_payloads(
+        self,
+    ) -> None:
+        """Cover GitHub payload parsing paths."""
         with patch.object(
             build_quality_rollup,
             "load_json_https",
@@ -125,7 +146,18 @@ class QualityRollupExtraTests(unittest.TestCase):
         ):
             payload = build_quality_rollup._github_payload("owner/repo", "sha", "token")
         self.assertIn("check_runs", payload)
+        with (
+            patch.object(
+                build_quality_rollup, "load_json_https", return_value=(["invalid"], {})
+            ),
+            self.assertRaisesRegex(
+                RuntimeError, "Unexpected GitHub API response payload"
+            ),
+        ):
+            build_quality_rollup._github_payload("owner/repo", "sha", "token")
 
+    def test_load_check_contexts_merges_status_payloads(self) -> None:
+        """Cover load_check_contexts with mixed response types."""
         responses = [
             {
                 "check_runs": [
@@ -148,18 +180,8 @@ class QualityRollupExtraTests(unittest.TestCase):
         self.assertEqual(contexts["DeepScan"]["source"], "status")
         self.assertNotIn("", contexts)
 
-        with (
-            patch.object(
-                build_quality_rollup, "load_json_https", return_value=(["invalid"], {})
-            ),
-            self.assertRaisesRegex(
-                RuntimeError, "Unexpected GitHub API response payload"
-            ),
-        ):
-            build_quality_rollup._github_payload("owner/repo", "sha", "token")
-
     def test_quality_rollup_main_passes_with_token_and_lane_payloads(self) -> None:
-        """Cover quality rollup main passes with token and lane payloads."""
+        """Cover the successful quality-rollup main path."""
         with tempfile.TemporaryDirectory() as temp_dir:
             root = Path(temp_dir)
             profile_path = root / "profile.json"
@@ -172,14 +194,7 @@ class QualityRollupExtraTests(unittest.TestCase):
                 ),
                 encoding="utf-8",
             )
-            args = Namespace(
-                profile_json=str(profile_path),
-                repo="owner/repo",
-                sha="abc123",
-                artifacts_dir=str(root),
-                out_json="quality-rollup/summary.json",
-                out_md="quality-rollup/summary.md",
-            )
+            args = self._rollup_main_args(root, profile_path)
             with (
                 patch.object(build_quality_rollup, "parse_args", return_value=args),
                 patch.object(
@@ -206,8 +221,8 @@ class QualityRollupExtraTests(unittest.TestCase):
                 self.assertEqual(build_quality_rollup.main(), 0)
         self.assertEqual(write_report_mock.call_args.args[0]["status"], "pass")
 
-    def test_build_rollup_and_wait_cover_pending_and_report_failure_paths(self) -> None:
-        """Cover build rollup and wait cover pending and report failure paths."""
+    def test_build_rollup_returns_pending_for_in_progress_contexts(self) -> None:
+        """Cover pending rollup status when required contexts are still running."""
         profile = {
             "slug": "owner/repo",
             "active_required_contexts": ["Coverage 100 Gate"],
@@ -224,12 +239,16 @@ class QualityRollupExtraTests(unittest.TestCase):
         )
         self.assertEqual(rollup["status"], "pending")
 
+    def test_quality_rollup_waits_for_pending_contexts(self) -> None:
+        """Cover wait_for_contexts using the shared helper sequence."""
         contexts, sleep_mock = exercise_wait_for_contexts(
             pending_then_success_contexts()
         )
         self.assertEqual(contexts["Coverage 100 Gate"]["conclusion"], "success")
         sleep_mock.assert_called_once()
 
+    def test_quality_rollup_main_propagates_write_report_failures(self) -> None:
+        """Cover main when report writing fails after a successful rollup."""
         with tempfile.TemporaryDirectory() as temp_dir:
             root = Path(temp_dir)
             profile_path = root / "profile.json"
